+++ tests/basic/helper.c
@@ -60,7 +60,7 @@ typedef struct td_sjs_string sjs_string;
 typedef struct td_sjs_class sjs_class;
 
 struct td_sjs_object {
-    intptr_t _refCount;
+    int _refCount;
 };
 
 struct td_sjs_interface {
